+++ codex-rs/exec-server/src/server.rs
@@ -1,10 +1,12 @@
 mod filesystem;
 mod handler;
+mod process_handler;
 mod processor;
 mod registry;
 mod transport;
 
 pub(crate) use handler::ExecServerHandler;
+pub(crate) use process_handler::ProcessHandler;
 pub use transport::DEFAULT_LISTEN_URL;
 pub use transport::ExecServerListenUrlParseError;
